requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==2.0.20
alembic==1.10.0
psycopg2-binary==2.9.6
pydantic==2.1.1
python-dotenv==1.0.0
pytest==7.4.0
pytest-cov==4.1.0
